Rename ProcessData to LoadData
